DataBlockHashIndex: remove the uint16 cast for map_offset

+++ table/block.cc
@@ -251,8 +251,7 @@ bool DataBlockIter::SeekForGet(const Slice& target) {
   }
 
   Slice user_key = ExtractUserKey(target);
-  uint16_t map_offset =
-      static_cast<uint16_t>(restarts_ + num_restarts_ * sizeof(uint32_t));
+  uint32_t map_offset = restarts_ + num_restarts_ * sizeof(uint32_t);
   uint8_t entry = data_block_hash_index_->Lookup(data_, map_offset, user_key);
 
   if (entry == kNoEntry) {
